+++ src/Layout/Header/Header.tsx
@@ -1,191 +1,52 @@
-import React, { useState, useEffect, MouseEvent } from "react";
-import { useNavigate, NavLink } from "react-router-dom";
-import {
-  AppBar,
-  Box,
-  Toolbar,
-  Tooltip,
-  IconButton,
-  ListItemIcon,
-  Typography,
-  Badge,
-  MenuItem,
-  Menu,
-  Avatar,
-  Divider,
-} from "@mui/material";
-
-interface Props {
-  handleDrawerOpen?: () => void;
-  open?: boolean;
-}
-
+import React, { useState } from "react";
+import { styled } from "@mui/material/styles";
+import { Box, Toolbar, Typography, IconButton, Badge } from "@mui/material";
+import MuiAppBar, { AppBarProps as MuiAppBarProps } from "@mui/material/AppBar";
 import MenuIcon from "@mui/icons-material/Menu";
 import AccountCircle from "@mui/icons-material/AccountCircle";
 import NotificationsIcon from "@mui/icons-material/Notifications";
-import MoreIcon from "@mui/icons-material/MoreVert";
-import Settings from "@mui/icons-material/Settings";
-import Logout from "@mui/icons-material/Logout";
-import AccountCircleOutlinedIcon from "@mui/icons-material/AccountCircleOutlined";
-import ManageAccountsOutlinedIcon from "@mui/icons-material/ManageAccountsOutlined";
 
-import { SearchBar } from "../../components";
+import { SearchBar } from "../../shared-components";
+import DrawerMenu from "./DrawerMenu";
+const drawerWidth = 240;
 
-const Header: React.FC<Props> = (props) => {
-  const { handleDrawerOpen, open } = props;
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-  const [mobileMoreAnchorEl, setMobileMoreAnchorEl] =
-    useState<null | HTMLElement>(null);
-  const navigate = useNavigate();
+interface AppBarProps extends MuiAppBarProps {
+  open?: boolean;
+}
 
-  const [isToken, setToken] = useState<string | null>();
-  const isMenuOpen = Boolean(anchorEl);
-  const isMobileMenuOpen = Boolean(mobileMoreAnchorEl);
-  const handleClick = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-  const handleProfileMenuOpen = (event: MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
+const AppBar = styled(MuiAppBar, {
+  shouldForwardProp: (prop) => prop !== "open",
+})<AppBarProps>(({ theme, open }) => ({
+  zIndex: theme.zIndex.drawer + 1,
+  transition: theme.transitions.create(["width", "margin"], {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen,
+  }),
+  ...(open && {
+    marginLeft: drawerWidth,
+    width: `calc(100% - ${drawerWidth}px)`,
+    transition: theme.transitions.create(["width", "margin"], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+  }),
+}));
 
-  const handleMobileMenuClose = () => {
-    setMobileMoreAnchorEl(null);
-  };
+const Header = () => {
+  const [open, setOpen] = useState<boolean>(false);
 
-  const handleMenuClose = () => {
-    setAnchorEl(null);
-    handleMobileMenuClose();
-  };
-  const logOut = () => {
-    localStorage.removeItem("token");
-    navigate("/login");
-    handleMenuClose();
+  const handleDrawerOpen = () => {
+    setOpen(true);
   };
 
-  const handleMobileMenuOpen = (event: MouseEvent<HTMLElement>) => {
-    setMobileMoreAnchorEl(event.currentTarget);
+  const handleDrawerClose = () => {
+    setOpen(false);
   };
-
-  useEffect(() => {
-    const token = localStorage.getItem("token");
-    setToken(token);
-  }, [isToken]);
-
   const menuId = "primary-search-account-menu";
 
-  const privateLinks = (
-    <div>
-      <MenuItem>
-        <ListItemIcon>
-          <AccountCircleOutlinedIcon fontSize="small" />
-        </ListItemIcon>{" "}
-        Profile
-      </MenuItem>
-      <MenuItem>
-        <ListItemIcon>
-          <ManageAccountsOutlinedIcon fontSize="small" />
-        </ListItemIcon>
-        My account
-      </MenuItem>
-      <Divider />
-      <MenuItem>
-        <ListItemIcon>
-          <Settings fontSize="small" />
-        </ListItemIcon>
-        Settings
-      </MenuItem>
-      <MenuItem onClick={logOut}>
-        <ListItemIcon>
-          <Logout fontSize="small" />
-        </ListItemIcon>
-        Logout
-      </MenuItem>
-    </div>
-  );
-  const renderMenu = (
-    <Menu
-      anchorEl={anchorEl}
-      id="account-menu"
-      open={isMenuOpen}
-      onClose={handleClose}
-      onClick={handleClose}
-      PaperProps={{
-        elevation: 0,
-        sx: {
-          overflow: "visible",
-          filter: "drop-shadow(0px 2px 5px rgba(0,0,0,0.20))",
-          width: 218,
-          mt: 1.5,
-          "& .MuiAvatar-root": {
-            width: 32,
-            height: 32,
-            ml: -0.5,
-            mr: 1,
-          },
-          "&:before": {
-            content: "''",
-            display: "block",
-            position: "absolute",
-            top: 0,
-            right: 14,
-            width: 10,
-            height: 10,
-            bgcolor: "background.paper",
-            transform: "translateY(-50%) rotate(45deg)",
-            zIndex: 0,
-          },
-        },
-      }}
-      transformOrigin={{ horizontal: "right", vertical: "top" }}
-      anchorOrigin={{ horizontal: "right", vertical: "bottom" }}
-    >
-      {!isToken && (
-        <MenuItem onClick={handleMenuClose}>
-          <NavLink to="/login"> Log in</NavLink>
-        </MenuItem>
-      )}
-      {isToken && privateLinks}
-    </Menu>
-  );
-
-  const mobileMenuId = "primary-search-account-menu-mobile";
-  const renderMobileMenu = (
-    <Menu
-      anchorEl={mobileMoreAnchorEl}
-      anchorOrigin={{
-        vertical: "top",
-        horizontal: "right",
-      }}
-      id={mobileMenuId}
-      keepMounted
-      transformOrigin={{
-        vertical: "top",
-        horizontal: "right",
-      }}
-      open={isMobileMenuOpen}
-      onClose={handleMobileMenuClose}
-    >
-      <Tooltip title="Account settings">
-        <IconButton
-          onClick={handleClick}
-          size="small"
-          sx={{ ml: 2 }}
-          aria-controls={isMenuOpen ? "account-menu" : undefined}
-          aria-haspopup="true"
-          aria-expanded={isMenuOpen ? "true" : undefined}
-        >
-          <Avatar sx={{ width: 32, height: 32 }}>M</Avatar>
-        </IconButton>
-      </Tooltip>
-    </Menu>
-  );
-
   return (
-    <Box sx={{ flexGrow: 1 }}>
-      <AppBar position="fixed">
+    <Box sx={{ display: "flex" }}>
+      <AppBar position="fixed" open={open}>
         <Toolbar>
           <IconButton
             color="inherit"
@@ -199,18 +60,11 @@ const Header: React.FC<Props> = (props) => {
           >
             <MenuIcon />
           </IconButton>
-          <Typography
-            variant="h6"
-            noWrap
-            component="div"
-            sx={{ display: { xs: "none", sm: "block" } }}
-          >
-            LOGO
+          <Typography variant="h6" noWrap component="div">
+            Mini variant drawer
           </Typography>
-
           <Box sx={{ flexGrow: 1 }} />
-
-          {/* <Box sx={{ display: { xs: "none", md: "flex" } }}>
+          <Box sx={{ display: { xs: "none", md: "flex" } }}>
             <SearchBar />
             <IconButton
               size="large"
@@ -227,30 +81,17 @@ const Header: React.FC<Props> = (props) => {
               aria-label="account of current user"
               aria-controls={menuId}
               aria-haspopup="true"
-              onClick={handleProfileMenuOpen}
+              // onClick={handleProfileMenuOpen}
               color="inherit"
             >
               <AccountCircle />
             </IconButton>
-          </Box> */}
-
-          <Box sx={{ display: { xs: "flex", md: "none" } }}>
-            <IconButton
-              size="large"
-              aria-label="show more"
-              aria-controls={mobileMenuId}
-              aria-haspopup="true"
-              onClick={handleMobileMenuOpen}
-              color="inherit"
-            >
-              <MoreIcon />
-            </IconButton>
           </Box>
         </Toolbar>
       </AppBar>
-      {renderMobileMenu}
-      {renderMenu}
+      <DrawerMenu open={open} handleDrawerClose={handleDrawerClose} />
     </Box>
   );
 };
+
 export default Header;
