+++ src/shared-components/ProfileAvatar/index.tsx
