+++ src/shared-components/dialogBox/style.tsx
