+++ src/shared-components/dialogBox/types.tsx
