+++ src/shared-components/searchBar/style.tsx
