+++ src/shared-components/vendorCard/types.tsx
